// types/staff-evaluation.ts
export interface StaffEvaluation {
  id: string
  hotelId: string
  staffFriendliness: number
  staffAppearance: number
  staffSmile: number
  staffTone: number
  staffToneReason: string
  staffComplaintHandling: number
  staffComplaintResolution: number
  managerHandling: number
  userId: string
  userName: string
  createdAt: string
}